+++ source/extensions/transport_sockets/tls/context_impl.cc
@@ -485,18 +485,18 @@ std::vector<Ssl::PrivateKeyMethodProviderSharedPtr> ContextImpl::getPrivateKeyMe
   return providers;
 }
 
-absl::optional<size_t> ContextImpl::daysUntilFirstCertExpires() const {
-  absl::optional<size_t> daysUntilExpiration = cert_validator_->daysUntilFirstCertExpires();
+absl::optional<uint32_t> ContextImpl::daysUntilFirstCertExpires() const {
+  absl::optional<uint32_t> daysUntilExpiration = cert_validator_->daysUntilFirstCertExpires();
   if (!daysUntilExpiration.has_value()) {
     return absl::nullopt;
   }
   for (auto& ctx : tls_contexts_) {
-    const absl::optional<size_t> tmp =
+    const absl::optional<uint32_t> tmp =
         Utility::getDaysUntilExpiration(ctx.cert_chain_.get(), time_source_);
     if (!tmp.has_value()) {
       return absl::nullopt;
     }
-    daysUntilExpiration = std::min<size_t>(tmp.value(), daysUntilExpiration.value());
+    daysUntilExpiration = std::min<uint32_t>(tmp.value(), daysUntilExpiration.value());
   }
   return daysUntilExpiration;
 }
